@@ -28,7 +28,7 @@ endif()
 INCLUDE(ExternalProject)
 INCLUDE(CheckCXXSourceCompiles)
 
-SET(CMAKE_CXX_STANDARD 17)
+SET(CMAKE_CXX_STANDARD 20)
 SET(CMAKE_CXX_STANDARD_REQUIRED TRUE)
 SET(CMAKE_CXX_EXTENSIONS FALSE)
 SET(CMAKE_EXPORT_COMPILE_COMMANDS TRUE)
@@ -214,6 +214,11 @@ IF (MASK_LONGDOUBLE)
     MY_CHECK_AND_SET_COMPILER_FLAG("-DMASK_LONGDOUBLE")
 ENDIF()
 
+if (CMAKE_CXX_COMPILER_ID STREQUAL "Clang")
+    MY_CHECK_AND_SET_COMPILER_FLAG("-Wno-deprecated-enum-enum-conversion -Wno-register")
+elseif (CMAKE_CXX_COMPILER_ID STREQUAL "GNU")
+    MY_CHECK_AND_SET_COMPILER_FLAG("-Wno-register")
+endif()
 
 MY_CHECK_AND_SET_COMPILER_FLAG("-Werror -Wall")
 SET (ENGINE_LDFLAGS    "-Wl,--no-as-needed -Wl,--add-needed")
