@@ -20,20 +20,17 @@
  */
 #include "MetadataFile.h"
 #include <boost/filesystem.hpp>
-#define BOOST_SPIRIT_THREADSAFE
-#include <boost/property_tree/ptree.hpp>
-#include <boost/property_tree/json_parser.hpp>
-#include <boost/foreach.hpp>
 #include <boost/uuid/uuid.hpp>
 #include <boost/uuid/uuid_io.hpp>
 #include <boost/uuid/random_generator.hpp>
+#include <cstdio>
 #include <unistd.h>
+#include <fstream>
 
 #define max(x, y) (x > y ? x : y)
 #define min(x, y) (x < y ? x : y)
 
 using namespace std;
-namespace bpt = boost::property_tree;
 namespace bf = boost::filesystem;
 
 namespace
@@ -120,12 +117,13 @@ MetadataFile::MetadataFile(const boost::filesystem::path& filename)
   {
     if (boost::filesystem::exists(mFilename))
     {
-      jsontree.reset(new bpt::ptree());
-      boost::property_tree::read_json(mFilename.string(), *jsontree);
+      std::ifstream i(mFilename.string());
+      jsontree.reset(new nlohmann::json);
+      i >> *jsontree;
       jsonCache.put(mFilename, jsontree);
       s.unlock();
       mVersion = 1;
-      mRevision = jsontree->get<int>("revision");
+      mRevision = (*jsontree)["revision"];
     }
     else
     {
@@ -140,7 +138,7 @@ MetadataFile::MetadataFile(const boost::filesystem::path& filename)
   {
     s.unlock();
     mVersion = 1;
-    mRevision = jsontree->get<int>("revision");
+    mRevision = (*jsontree)["revision"];;
   }
   ++metadataFilesAccessed;
 }
@@ -162,12 +160,13 @@ MetadataFile::MetadataFile(const boost::filesystem::path& filename, no_create_t,
     if (boost::filesystem::exists(mFilename))
     {
       _exists = true;
-      jsontree.reset(new bpt::ptree());
-      boost::property_tree::read_json(mFilename.string(), *jsontree);
+      jsontree.reset(new nlohmann::json);
+      std::ifstream i(mFilename.string());
+      i >> *jsontree;
       jsonCache.put(mFilename, jsontree);
       s.unlock();
       mVersion = 1;
-      mRevision = jsontree->get<int>("revision");
+      mRevision = (*jsontree)["revision"];
     }
     else
     {
@@ -182,7 +181,7 @@ MetadataFile::MetadataFile(const boost::filesystem::path& filename, no_create_t,
     s.unlock();
     _exists = true;
     mVersion = 1;
-    mRevision = jsontree->get<int>("revision");
+    mRevision = (*jsontree)["revision"];
   }
   ++metadataFilesAccessed;
 }
@@ -193,11 +192,10 @@ MetadataFile::~MetadataFile()
 
 void MetadataFile::makeEmptyJsonTree()
 {
-  jsontree.reset(new bpt::ptree());
-  boost::property_tree::ptree objs;
-  jsontree->put("version", mVersion);
-  jsontree->put("revision", mRevision);
-  jsontree->add_child("objects", objs);
+  jsontree.reset(new nlohmann::json);
+  (*jsontree)["version"] = mVersion;
+  (*jsontree)["revision"] = mRevision;
+  (*jsontree)["objects"] = nlohmann::json::array();
 }
 
 void MetadataFile::printKPIs()
@@ -219,11 +217,11 @@ size_t MetadataFile::getLength() const
 {
   size_t totalSize = 0;
 
-  auto& objects = jsontree->get_child("objects");
+  auto &objects = (*jsontree)["objects"];
   if (!objects.empty())
   {
-    auto& lastObject = objects.back().second;
-    totalSize = lastObject.get<off_t>("offset") + lastObject.get<size_t>("length");
+    auto& lastObject = objects.back();
+    totalSize = lastObject["offset"].get<off_t>() + lastObject["length"].get<size_t>();
   }
   return totalSize;
 }
@@ -243,10 +241,9 @@ vector<metadataObject> MetadataFile::metadataRead(off_t offset, size_t length) c
      rather than write a new alg.
   */
   set<metadataObject> mObjects;
-  BOOST_FOREACH (const boost::property_tree::ptree::value_type& v, jsontree->get_child("objects"))
+  for(const auto &v : (*jsontree)["objects"])
   {
-    mObjects.insert(metadataObject(v.second.get<uint64_t>("offset"), v.second.get<uint64_t>("length"),
-                                   v.second.get<string>("key")));
+    mObjects.insert(metadataObject(v["offset"], v["length"], v["key"]));
   }
 
   if (mObjects.size() == 0)
@@ -288,20 +285,20 @@ metadataObject MetadataFile::addMetadataObject(const boost::filesystem::path& fi
   //
 
   metadataObject addObject;
-  auto& objects = jsontree->get_child("objects");
+  auto& objects = (*jsontree)["objects"];
   if (!objects.empty())
   {
-    auto& lastObject = objects.back().second;
-    addObject.offset = lastObject.get<off_t>("offset") + mpConfig->mObjectSize;
+    auto& lastObject = objects.back();
+    addObject.offset = lastObject["offset"].get<off_t>() + mpConfig->mObjectSize;
   }
 
   addObject.length = length;
   addObject.key = getNewKey(filename.string(), addObject.offset, addObject.length);
-  boost::property_tree::ptree object;
-  object.put("offset", addObject.offset);
-  object.put("length", addObject.length);
-  object.put("key", addObject.key);
-  objects.push_back(make_pair("", object));
+  nlohmann::json object = nlohmann::json::object();
+  object["offset"] = addObject.offset;
+  object["length"] = addObject.length;
+  object["key"] = addObject.key;
+  objects.push_back(object);
 
   return addObject;
 }
@@ -312,7 +309,8 @@ int MetadataFile::writeMetadata()
   if (!boost::filesystem::exists(mFilename.parent_path()))
     boost::filesystem::create_directories(mFilename.parent_path());
 
-  write_json(mFilename.string(), *jsontree);
+  std::ofstream o(mFilename.c_str());
+  o << *jsontree;
   _exists = true;
 
   boost::unique_lock<boost::mutex> s(jsonCache.getMutex());
@@ -324,13 +322,13 @@ int MetadataFile::writeMetadata()
 bool MetadataFile::getEntry(off_t offset, metadataObject* out) const
 {
   metadataObject addObject;
-  BOOST_FOREACH (const boost::property_tree::ptree::value_type& v, jsontree->get_child("objects"))
+  for(auto &v: (*jsontree)["objects"])
   {
-    if (v.second.get<off_t>("offset") == offset)
+    if (v["offset"].get<off_t>() == offset)
     {
       out->offset = offset;
-      out->length = v.second.get<size_t>("length");
-      out->key = v.second.get<string>("key");
+      out->length = v["length"].get<size_t>();
+      out->key = v["key"];
       return true;
     }
   }
@@ -339,10 +337,10 @@ bool MetadataFile::getEntry(off_t offset, metadataObject* out) const
 
 void MetadataFile::removeEntry(off_t offset)
 {
-  bpt::ptree& objects = jsontree->get_child("objects");
-  for (bpt::ptree::iterator it = objects.begin(); it != objects.end(); ++it)
+  auto& objects = (*jsontree)["objects"];
+  for (auto it = objects.begin(); it != objects.end(); ++it)
   {
-    if (it->second.get<off_t>("offset") == offset)
+    if ((*it)["offset"].get<off_t>() == offset)
     {
       objects.erase(it);
       break;
@@ -352,7 +350,7 @@ void MetadataFile::removeEntry(off_t offset)
 
 void MetadataFile::removeAllEntries()
 {
-  jsontree->get_child("objects").clear();
+  (*jsontree)["objects"] = nlohmann::json::array();
 }
 
 void MetadataFile::deletedMeta(const bf::path& p)
@@ -456,21 +454,21 @@ void MetadataFile::setLengthInKey(string& key, size_t newLength)
 
 void MetadataFile::printObjects() const
 {
-  BOOST_FOREACH (const boost::property_tree::ptree::value_type& v, jsontree->get_child("objects"))
+  for (auto& v : (*jsontree)["objects"])
   {
-    printf("Name: %s Length: %zu Offset: %lld\n", v.second.get<string>("key").c_str(),
-           v.second.get<size_t>("length"), (long long)v.second.get<off_t>("offset"));
+    printf("Name: %s Length: %zu Offset: %lld\n", v["key"].get<std::string>().c_str(),
+           v["length"].get<size_t>(), (long long)v["offset"].get<off_t>());
   }
 }
 
 void MetadataFile::updateEntry(off_t offset, const string& newName, size_t newLength)
 {
-  for (auto& v : jsontree->get_child("objects"))
+  for (auto& v : (*jsontree)["objects"])
   {
-    if (v.second.get<off_t>("offset") == offset)
+    if (v["offset"].get<off_t>() == offset)
     {
-      v.second.put("key", newName);
-      v.second.put("length", newLength);
+      v["key"] = newName;
+      v["length"] = newLength;
       return;
     }
   }
@@ -482,11 +480,11 @@ void MetadataFile::updateEntry(off_t offset, const string& newName, size_t newLe
 
 void MetadataFile::updateEntryLength(off_t offset, size_t newLength)
 {
-  for (auto& v : jsontree->get_child("objects"))
+  for (auto& v : (*jsontree)["objects"])
   {
-    if (v.second.get<off_t>("offset") == offset)
+    if (v["offset"].get<off_t>() == offset)
     {
-      v.second.put("length", newLength);
+      v["length"] = newLength;
       return;
     }
   }
@@ -498,11 +496,12 @@ void MetadataFile::updateEntryLength(off_t offset, size_t newLength)
 
 off_t MetadataFile::getMetadataNewObjectOffset()
 {
-  auto& objects = jsontree->get_child("objects");
+  auto& objects = (*jsontree)["objects"];
   if (objects.empty())
     return 0;
-  auto& lastObject = jsontree->get_child("objects").back().second;
-  return lastObject.get<off_t>("offset") + lastObject.get<size_t>("length");
+
+  auto& lastObject = objects.back();
+  return lastObject["offset"].get<off_t>() + lastObject["length"].get<size_t>();
 }
 
 metadataObject::metadataObject() : offset(0), length(0)
