@@ -479,7 +479,7 @@ class JobStep
   bool fDelivery;
   bool fOnClauseFilter;
   volatile bool fDie;
-  volatile uint32_t fWaitToRunStepCnt;
+  uint32_t fWaitToRunStepCnt;
   std::string fExtendedInfo;
   std::string fMiniInfo;
 
