@@ -40,7 +40,7 @@
 #ifdef POSIX_REGEX
 #include <regex.h>
 #else
-#include <boost/regex.hpp>
+#include <regex>
 #endif
 #include <cstddef>
 #include <boost/shared_ptr.hpp>
