@@ -28,7 +28,7 @@ using namespace std;
 #ifdef __linux__
 #include <regex.h>
 #else
-#include <boost/regex.hpp>
+#include <regex>
 using namespace boost;
 #endif
 
@@ -226,8 +226,8 @@ inline bool getBool(rowgroup::Row& row, funcexp::FunctionParm& pm, bool& isNull,
     return false;
 
 #else
-  regex pat(pattern.c_str());
-  return regex_search(expr.c_str(), pat);
+  std::regex pat(pattern.c_str());
+  return std::regex_search(expr.c_str(), pat);
 #endif
 }
 
